+++ LoginServer/src/main/java/com/redhat/login/mapper/UserMapper.java
@@ -10,23 +10,23 @@ import org.apache.ibatis.annotations.Param;
 public interface UserMapper {
 
 	// 获取用户id
-	public int getId(String username);
+	public long getId(String username);
 
 	// 获取用户密码
-	public String getPassword(int id);
+	public String getPassword(long id);
 
 	// 添加用户
 	public boolean addUser(@Param("username") String username, @Param("password") String password);
 
 	// 更新用户登录时间
-	public boolean updateOnline(@Param("id") int id, @Param("date") Date date);
+	public boolean updateOnline(@Param("id") long id, @Param("date") Date date);
 
 	// 获取用户登录时间
-	public Date getOnline(int id);
+	public Date getOnline(long id);
 
 	// 更新用户登录时间
-	public boolean updateEnable(@Param("id") int id, @Param("enable") boolean enable);
+	public boolean updateEnable(@Param("id") long id, @Param("enable") boolean enable);
 
 	// 获取是否允许登录
-	public boolean getEnable(int id);
+	public boolean getEnable(long id);
 }
